style: 태그 변경 (#ATR-18)
