+++ apps/service/entities/profile/index.ts
@@ -1,2 +1,2 @@
-export * as Profile from './ui'
+export * from './ui'
 export * from './model'
