+++ apps/service/features/calendar/model/index.ts
@@ -1,0 +1,1 @@
+export * from './types'
