+++ apps/service/features/graph/ui/index.tsx
@@ -1,11 +1,1 @@
-import { ToolEmoji } from '@attraction/icons'
-import React from 'react'
-
-export default function Graph() {
-  return (
-    <div className="flex w-full flex-col items-center justify-center rounded-2xl border border-gray-100 bg-white p-4">
-      <ToolEmoji className="size-16" />
-      <div className="mt-5 font-semibold">아직 준비중이에요 기다려주세요</div>
-    </div>
-  )
-}
+export { default as Graph } from './Graph'
