--- apps/service/features/index.ts
@@ -1,5 +1,0 @@
-export * from './profile'
-export * from './calendar'
-export * from './graph'
-export * from './recent-newsletter'
-export * from './subscribe-list'
