+++ apps/service/features/profile-container/index.ts
@@ -1,3 +1,2 @@
 export * from './ui'
 export * from './api'
-export * from './libs'
