+++ apps/service/features/profile-container/ui/index.tsx
@@ -1,0 +1,1 @@
+export { default as ProfileContainer } from './ProfileContainer'
