--- apps/service/features/profile/index.ts
@@ -1,2 +1,0 @@
-export { default as Profiles } from './ui'
-export * from './api'
