+++ apps/service/features/recent-newsletter/index.ts
@@ -1,2 +1,2 @@
-export { default as RecentNewsletter } from './ui'
 export * from './api'
+export * from './ui'
