+++ apps/service/features/subscribe-list/index.ts
@@ -1,2 +1,2 @@
-export { default as SubscribeList } from './ui'
 export * from './api'
+export * from './ui'
