--- apps/service/shared/image-box/index.ts
@@ -1,1 +1,0 @@
-export { default as ImageBox } from './ui'
