+++ apps/service/shared/ui/content/index.ts
