+++ apps/service/shared/ui/image-box/index.ts
