+++ apps/service/shared/ui/image-box/ui/ImageBox.tsx
