+++ apps/service/shared/ui/image-box/ui/index.tsx
@@ -1,0 +1,1 @@
+export { default as ImageBox } from './ImageBox'
